Форма создания поста вынесена на страницу списка постов - CRLF -> LF - Сделал маршруты более понятными - Избавился от ненужного контроллера, как мне кажется, перенес его поведение в ФорумКонтроллер - Поправил тесты
