--- src/ForumBundle/Controller/DefaultController.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace ForumBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-
-
-class DefaultController extends Controller
-{
-    /**
-     * Lists all Forum entities.
-     */
-    public function indexAction()
-    {
-        $em = $this->getDoctrine()->getManager();
-
-        $forums = $em->getRepository('ForumBundle:Forum')->findAll();
-
-        return $this->render('@Forum/forum/index.html.twig', [
-            'forums' => $forums,
-        ]);
-    }
-}
