+++ src/ForumBundle/Controller/ForumController.php
@@ -3,12 +3,24 @@
 namespace ForumBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
 use ForumBundle\Entity\Forum;
-
 
 class ForumController extends Controller
 {
+    /**
+     * Lists all Forum entities.
+     */
+    public function indexAction()
+    {
+        $em = $this->getDoctrine()->getManager();
+
+        $forums = $em->getRepository('ForumBundle:Forum')->findAll();
+
+        return $this->render('@Forum/forum/index.html.twig', [
+            'forums' => $forums,
+        ]);
+    }
+
     /**
      * Список топиков в форуме
      *
@@ -16,7 +28,7 @@ class ForumController extends Controller
      * @param int $page
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function indexAction(Forum $forum, $page)
+    public function showAction(Forum $forum, $page)
     {
         $q = $this->getDoctrine()->getRepository('ForumBundle:Topic')->getListQuery($forum);
