+++ src/ForumBundle/Controller/PostController.php
@@ -4,10 +4,8 @@ namespace ForumBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
-
 use ForumBundle\Entity\Topic;
 use ForumBundle\Entity\Post;
-
 use ForumBundle\Form\PostType;
 
 /**
@@ -23,11 +21,12 @@ class PostController extends Controller
      * @param Topic $topic
      * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
      */
-    public function newAction(Request $request, Topic $topic) {
-        $form = $this->createForm(PostType::class);
-        $form->handleRequest($request);
-        $post = $form->getData();
+    public function newAction(Request $request, Topic $topic)
+    {
+        $post = new Post();
         $post->setTopic($topic);
+        $form = $this->createForm(PostType::class, $post);
+        $form->handleRequest($request);
         if ($form->isSubmitted()) {
             if ($form->isValid()) { //TODO: catch error
                 $em = $this->getDoctrine()->getManager();
@@ -36,7 +35,7 @@ class PostController extends Controller
             }
         }
 
-        return $this->redirectToRoute('topic_show', ['id' => 1]);
+        return $this->redirectToRoute('topic_show', ['id' => $topic->getId()]);
     }
 
     /**
