+++ src/ForumBundle/Controller/TopicController.php
@@ -4,14 +4,11 @@ namespace ForumBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
-
 use ForumBundle\Form\TopicType;
 use ForumBundle\Form\PostType;
-
 use ForumBundle\Entity\Forum;
 use ForumBundle\Entity\Topic;
 use ForumBundle\Entity\Post;
-
 
 class TopicController extends Controller
 {
@@ -22,11 +19,12 @@ class TopicController extends Controller
      * @param int $page
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function indexAction(Topic $topic, $page)
+    public function showAction(Topic $topic, $page)
     {
         $q = $this->getDoctrine()->getRepository('ForumBundle:Post')->getListQuery($topic);
 
         $pager = $this->get('paginate')->paginate($q, $page);
+
         $form = $this->createForm(PostType::class, null, [
             'action' => $this->generateUrl('post_new', [
                 'id' => $topic->getId(),
@@ -36,7 +34,6 @@ class TopicController extends Controller
         return $this->render('@Forum/post/index.html.twig', [
             'topic' => $topic,
             'posts' => $pager,
-
             'form' => $form->createView(),
         ]);
     }
