+++ src/ForumBundle/Form/PostType.php
@@ -3,11 +3,10 @@
 namespace ForumBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-
 use ForumBundle\Entity\Post;
 
 class PostType extends AbstractType
@@ -19,7 +18,7 @@ class PostType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('text', TextType::class, ['label' => 'Ответ'])
+            ->add('text', TextareaType::class, ['label' => false])
             ->add('submit', SubmitType::class, ['label' => 'Отправить'])
         ;
     }
