+++ src/ForumBundle/Form/TopicType.php
@@ -3,12 +3,10 @@
 namespace ForumBundle\Form;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
-
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-
 use ForumBundle\Entity\Topic;
 
 class TopicType extends AbstractType
