+++ src/ForumBundle/Helper/Paginate.php
@@ -11,7 +11,6 @@ use Pagerfanta\Adapter\ArrayAdapter;
 use Pagerfanta\Pagerfanta;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
-
 class Paginate
 {
     /**
