+++ src/ForumBundle/Resources/config/routing/forum.yml
@@ -1,6 +1,6 @@
 forum_show:
-    path:     /{id}/{page}
-    defaults: { _controller: "ForumBundle:Forum:index", page: 1 }
+    path:     /{id}/show/{page}
+    defaults: { _controller: "ForumBundle:Forum:show", page: 1 }
     requirements:
         id:   \d+
         page: \d+
