+++ src/ForumBundle/Resources/views/topic/new.html.twig
@@ -5,6 +5,7 @@
 {% block content %}
     {{ form_start(form) }}
         {{ form_widget(form) }}
+    {# TODO: Добавить поле/textarea для сообщения/post #}
     {{ form_end(form) }}
     <a class="ui-btn ui-btn-inline ui-corner-all ui-shadow ui-icon-arrow-l ui-btn-icon-left" href="{{ path('forum_show', {'id': topic.forum.id}) }}">Вернуться</a>
 {% endblock %}
