+++ tests/ForumBundle/Controller/ForumControllerTest.php
@@ -7,14 +7,16 @@ use Symfony\Component\HttpFoundation\Response;
 
 class ForumControllerTest extends ForumWebTestCase
 {
-    public function testForum()
+    public function testIndex()
     {
-        $uri = self::$kernel->getContainer()->get('router')->generate('forum_show', ['id' => 1, 'page' => PHP_INT_MAX]);
+        $text = ['PHP', 'MySQL'];
+        $uri = self::$kernel->getContainer()->get('router')->generate('index');
 
         $crawler = self::$client->request('GET', $uri);
 
         $this->assertEquals(Response::HTTP_OK, self::$client->getResponse()->getStatusCode());
 
-        $this->assertContains('Test topic 1 in PHP forum', $crawler->filter('li > a')->last()->text());
+        $this->assertContains($text[0], $crawler->filter('li > a')->eq(0)->text());
+        $this->assertContains($text[1], $crawler->filter('li > a')->eq(1)->text());
     }
 }
